fix bug for created_by_top_10_contributor_count_first_changeset_monthly
This was a bug. The script was only looking at all changeset from the top 100 editors to determine the first used editor. If someone would have used an editor not in the top 100 and then use one from the top 100, it would have counted as "the first editor". This is now fixed.